【主题配置】添加标签栏风格 ：谷歌风格、按钮风格
