+++ src/language/locales/en.json
@@ -60,6 +60,7 @@
         "Auto close settings center",
         "Global watermark",
         "Menu width",
+        "Tab Mode",
         "Page animation",
         "Custom radius"
       ]
