+++ src/language/locales/zh.json
@@ -60,6 +60,7 @@
         "自动关闭设置中心",
         "全局水印",
         "菜单宽度",
+        "标签栏风格",
         "页面切换动画",
         "自定义圆角"
       ]
